wip: refactor fields and meta
