+++ web/src/components/molecules/Content/Form/fields/FieldComponents/BoolField.tsx
@@ -1,0 +1,25 @@
+import Form from "@reearth-cms/components/atoms/Form";
+import Switch from "@reearth-cms/components/atoms/Switch";
+import MultiValueBooleanField from "@reearth-cms/components/molecules/Common/MultiValueField/MultiValueBooleanField";
+import { Field } from "@reearth-cms/components/molecules/Schema/types";
+
+import FieldTitle from "../../FieldTitle";
+
+interface BoolFieldProps {
+  field: Field;
+  handleBlurUpdate?: () => void;
+}
+
+const BoolField: React.FC<BoolFieldProps> = ({ field }) => {
+  return (
+    <Form.Item
+      extra={field.description}
+      name={field.id}
+      valuePropName="checked"
+      label={<FieldTitle title={field.title} isUnique={field.unique} isTitle={field.isTitle} />}>
+      {field.multiple ? <MultiValueBooleanField FieldInput={Switch} /> : <Switch />}
+    </Form.Item>
+  );
+};
+
+export default BoolField;
